feat: store progress in ViewModel
