+++ app/src/main/java/app/revanced/manager/ui/component/ArrowButton.kt
@@ -13,30 +13,25 @@ import androidx.compose.ui.res.stringResource
 import app.revanced.manager.R
 
 @Composable
-fun ArrowButton(modifier: Modifier = Modifier, expanded: Boolean, onClick: () -> Unit) {
-    IconButton(onClick = onClick) {
-        val description = if (expanded) R.string.collapse_content else R.string.expand_content
-        val rotation by animateFloatAsState(targetValue = if (expanded) 0f else 180f, label = "rotation")
-
-        Icon(
-            imageVector = Icons.Filled.KeyboardArrowUp,
-            contentDescription = stringResource(description),
-            modifier = Modifier
-                .rotate(rotation)
-                .then(modifier)
-        )
-    }
-}
-
-@Composable
-fun Arrow(modifier: Modifier = Modifier, expanded: Boolean) {
+fun ArrowButton(modifier: Modifier = Modifier, expanded: Boolean, onClick: (() -> Unit)?) {
     val description = if (expanded) R.string.collapse_content else R.string.expand_content
     val rotation by animateFloatAsState(
         targetValue = if (expanded) 0f else 180f,
         label = "rotation"
     )
 
-    Icon(
+    onClick?.let {
+        IconButton(onClick = it) {
+
+            Icon(
+                imageVector = Icons.Filled.KeyboardArrowUp,
+                contentDescription = stringResource(description),
+                modifier = Modifier
+                    .rotate(rotation)
+                    .then(modifier)
+            )
+        }
+    } ?: Icon(
         imageVector = Icons.Filled.KeyboardArrowUp,
         contentDescription = stringResource(description),
         modifier = Modifier
